MINIFICPP-1659 Extract common blob storage functionality
